[1.3.2] brender/v1db/prepmesh: don't clear prepared data in BrModelClear()
"prepared" data is a v11model_f (and related) pointer, NOT the
device-specific data pointer, so this is safe. It will be cleared by
BrModelUpdate() and BrModelFree() in due course. Note that the prepared
pointer is BrResAllocate()'d with the model itself as the parent.

To explicitly clear it call BrModelClearPrepared().