[1.3.2] brender/{fw,glrend,inc,v1db}: update internal model format
BRender 1.3.2 switch from using an array of structures (interleaved)
to a struct of arrays (noninterleaved) to store vertex and face data.
Update the structures and preparation code, as well as glrend to handle
this.

Note that the glrend changes are the bare minimum to get it working with
the new format -- it still internally uses interleaved data. This should
be changed to match noninterleaved format that the engine gives it.