brender/glrend/outfcty: remove SDL screen mode enumeration
An "output facility"'s sole purpose is to create the "screen", or "front
buffer" in OpenGL nomenclature. This is already done before we enter the
driver, and one is created on-the-fly as needed, so the mode enumeration
isn't needed anymore.

This does mean that BrOutputFacilityEnum() can't be used properly with
glrend, but one can argue that dealing with screen modes isn't OpenGL's
problem.