brender: glrend: refactor device and screen creation
Refactors the driver such that all the OpenGL state is stored on the
screen pixelmap instead of the device object. Thus any OpenGL
initialisation is deferred until the screen is created, allowing the
device to be registered without a br_device_gl_ext procs structure.

Adds a new device pixelmap implementation to represent the screen
pixelmap (devpmglf). This solves the problem of moving the OpenGL
initialisation logic to the existing implementation making it even more
unwieldy than it already is. Additionally, switch to using the
br_device_pixelmap_fail implementations for most functionality -
the screen pixelmap is a special-case, and most pixelmap operations
should fail on it anyway.

Only one screen pixelmap can be created at a time, and the driver will
enforce this.

Also updates all the sample applications and tools to use the new
initialisation.

Casulities & Injuries:
- Casualty - MSAA - there's no way to pass arbitrary tokens via
  BrPixelmapMatch(), so an MSAA level cannot be set.
  I am unsure of how to work around this. Ideally, I'd like the
  screen pixelmap to be non-antialiased, only have it enabled on
  framebuffers by user request.
- Injury - Code cleanliness
  Most of the existing code relied on having easy access to the VIDEO
  structure and friends, via their "device" method. This is, of course,
  no longer possible due to everything now being on the screen pixelmap.

  Because only one screen pixelmap is allowed at a time, it can be
  obtained via the device structure. As a stopgap, a new helper method,
  DeviceGLGetScreen(), has been introduced for this purpose.

  The rest of the code should be refactored now that the original
  assumption is no longer valid. This will be done in a later patch.