core/math/fixed: convert BrFixedATan2() to not use float intermediaries
Note the parameter name change from (x, y) to (y, x) - The TRM says
BrFixedATan2(x, y) is equivalent to atan2f(x, y). The args to atan2f()
are (y, x), and the disparity is just confusing and annoying.

Oh, and as an added bonus, the inline documentation in the original ASM
had x/y flipped too.

Fix this mess by standardising on (y, x).