drivers/glrend: fix transparency sorting
Since the beginning of glrend, transparency sorting has been broken.
Object groups (i.e. material groups) are sorted by the distance from
their center to the camera and rendered appropriately.

This works for everything except transparent geometry. For Croc, this
is 99% okay, except in Dante's boss level, in which the stained-glass
windows are obviously broken.

To fix this, if a transparent group is encountered, it is broken-up
into individual triangles, each being sorted by its centroid, and then
rendered from back-to-front.

This significantly increases the no. of GL calls (because for each
triangle is effectively 3x groups), but this can be optimised
after-the-fact.