drivers/glrend/v1model: disable depth testing if no depth-buffer
When doing a ZS render, the BRT_DEPTH_BUFFER_O token is NULL,
but on the OpenGL-side there may actually be a depth buffer
attached.

If this is the case, applications won't clear the buffer (because
why would they?) and OpenGL is testing against an empty buffer
(i.e. everything's occluded).