core/fmt/savegltf: don't create textures for materials without a pixelmap
Originally I did a 1:1 mapping, and just had the texture not reference
an image. This, however, is undefined behaviour.

If a texture is present, it uses:
  <tex_colour> * <base_colour> * <colour attribute>

In Blender's case, a texture without an image is sampled as (0, 0, 0, 0)
causing the result to end up black, which is not what we want - we want
the material colour.

So remove it.