core/pixelmap: move clone functionality to match()
The original BRender function, BrPixelmapClone(), dispatched to match()
with use_type == BRT_CLONE. The newer BrPixelmapCloneTyped(), didn't.

To maintain consistency, do the same - move the current
BrPixelmapCloneTyped() to a pixelmap-resident function, and redirect
br_device_pixelmap_mem's BRT_CLONE handling (of which there was none) to
it.

Note that this breaks usage of BrPixelmapCloneTyped() with the SDL
drivers (and glrend, but that's never been supported). This will be
fixed in a followup commit.