{core,drivers/glrend}: resurrect BR_MATF_BLEND, don't blend if alpha-channel exists
It was used at some point in the past apparently, before 1.2, then
removed.

Restore it as there's currently no way to say "I want blending" other
than:
1. Have material opacity < 255
2. Have an alpha channel in the attached pixemap
3. Have the PREALPHA flag, which has different semantics.

(2) is a royal pain, especially when loading PNGs, which may be purely
opaque and have an alpha channel - we don't want these blended.

Having an explicit "I want blending" flag will allow us to remove (2)
with minimal fallout:
- It won't affect the 8-bit software renderer, as this is all done with
  blend tables.
- It _may_ affect the 15/16-bit software renderers, but literally no one
  uses these, and opacity is implemented via the screendoor renderer.
- The 3dfx/RAGE drivers don't exist anymore; their behaviour is
  irrelevant.

Note that we can't use the presence of the BR_MATM_BLEND_MODE_* values
to determine blending, as STANDARD is treated the same as DIMMED, and
STANDARD is the zero-value.